add react-ts-esbuild playground
