+++ playground/error-trackings/react-ts-esbuild/.gitignore
@@ -22,6 +22,3 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-
-# Sentry Config File
-.sentryclirc
