+++ playground/error-trackings/react-ts-esbuild/eslint.config.js
@@ -2,29 +2,24 @@ import js from '@eslint/js'
 import globals from 'globals'
 import reactHooks from 'eslint-plugin-react-hooks'
 import reactRefresh from 'eslint-plugin-react-refresh'
+import tseslint from 'typescript-eslint'
 
-export default [
+export default tseslint.config(
     { ignores: ['dist'] },
     {
-        files: ['**/*.{js,jsx}'],
+        extends: [js.configs.recommended, ...tseslint.configs.recommended],
+        files: ['**/*.{ts,tsx}'],
         languageOptions: {
             ecmaVersion: 2020,
             globals: globals.browser,
-            parserOptions: {
-                ecmaVersion: 'latest',
-                ecmaFeatures: { jsx: true },
-                sourceType: 'module',
-            },
         },
         plugins: {
             'react-hooks': reactHooks,
             'react-refresh': reactRefresh,
         },
         rules: {
-            ...js.configs.recommended.rules,
             ...reactHooks.configs.recommended.rules,
-            'no-unused-vars': ['error', { varsIgnorePattern: '^[A-Z_]' }],
             'react-refresh/only-export-components': ['warn', { allowConstantExport: true }],
         },
-    },
-]
+    }
+)
