+++ playground/error-trackings/react-ts-esbuild/package.json
@@ -1,0 +1,38 @@
+{
+    "name": "react-ts-esbuild",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+        "dev": "vite",
+        "build": "tsc -b && vite build",
+        "lint": "eslint .",
+        "inject": "posthog-cli sourcemap inject --directory ./dist/assets",
+        "preview": "vite preview",
+        "link-posthog": "pnpm link ../../..",
+        "build-posthog": "cd ../../.. && NODE_ENV=dev pnpm i && pnpm run build"
+    },
+    "dependencies": {
+        "posthog-js": "link:../../..",
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0"
+    },
+    "devDependencies": {
+        "@eslint/js": "^9.21.0",
+        "@types/react": "^19.0.10",
+        "@types/react-dom": "^19.0.4",
+        "@vitejs/plugin-react": "^4.3.4",
+        "eslint": "^9.21.0",
+        "eslint-plugin-react-hooks": "^5.1.0",
+        "eslint-plugin-react-refresh": "^0.4.19",
+        "globals": "^15.15.0",
+        "typescript": "~5.7.2",
+        "typescript-eslint": "^8.24.1",
+        "vite": "^6.2.0"
+    },
+    "pnpm": {
+        "overrides": {
+            "posthog-js": "link:../../.."
+        }
+    }
+}
