+++ playground/error-trackings/react-ts-esbuild/tsconfig.json
