+++ playground/error-trackings/react-ts-esbuild/tsconfig.node.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
         "target": "ES2022",
         "lib": ["ES2023"],
         "module": "ESNext",
