+++ playground/error-trackings/react-ts-esbuild/vite.config.ts
@@ -6,5 +6,6 @@ export default defineConfig({
     plugins: [react()],
     build: {
         sourcemap: true,
+        minify: 'esbuild',
     },
 })
