--- playground/sourcemaps/package.json
@@ -1,43 +1,0 @@
-{
-    "name": "sourcemaps",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "dev": "pnpm run link-posthog-js && vite",
-        "build": "vite build",
-        "start": "pnpm run build && posthog-cli sourcemap inject --directory ./dist/assets/ && pnpm run preview",
-        "lint": "eslint .",
-        "preview": "vite preview",
-        "build-posthog-js": "cd ../../ && NODE_ENV=dev pnpm i && pnpm run build",
-        "link-posthog-js": "pnpm link ../..",
-        "clean-react": "cd ../../react && pnpm run clean-node-modules"
-    },
-    "dependencies": {
-        "@sentry/cli": "^2.42.2",
-        "@types/node": "^22.13.9",
-        "posthog-js": "^1.215.7",
-        "react": "^19.0.0",
-        "react-dom": "^19.0.0",
-        "typescript": "^5.8.2"
-    },
-    "devDependencies": {
-        "@eslint/js": "^9.21.0",
-        "@sentry/cli": "^2.42.2",
-        "react": "^19.0.0",
-        "react-dom": "^19.0.0",
-        "@types/react": "^19.0.10",
-        "@types/react-dom": "^19.0.4",
-        "@vitejs/plugin-react": "^4.3.4",
-        "eslint": "^9.21.0",
-        "eslint-plugin-react-hooks": "^5.1.0",
-        "eslint-plugin-react-refresh": "^0.4.19",
-        "globals": "^15.15.0",
-        "vite": "^6.2.0"
-    },
-    "pnpm": {
-        "overrides": {
-            "posthog-js": "link:../.."
-        }
-    }
-}
